+++ src/solver.cpp
@@ -109,7 +109,7 @@ void FRSolver::setup(_mpi_comm comm_in, _mpi_comm comm_world)
   /* Create eles objects */
   unsigned int elesObjID = 0;
   ele2elesObj.assign({geo.nEles});
-  geo.ele_types.assign({geo.ele_set.size()});
+  geo.ele_types.assign({(uint)geo.ele_set.size()});
   if (input->iterative_method == MCGS)
   {
     for (auto etype : geo.ele_set)
